Fix test and fake repository
After changing the business logic, where months that already passed are shown as available false, tests started to fail. Those tests were fixed.

After changing the user token to be deleted once the user update its password, tests started to fail because the new method of delete token was not implemented in the fake repository used in the tests. This was fixed.