Apply class transformer to data saved on cache
Until now, we were saving the information on cache, without serializing it, leading to have unwanted fields.
This behavior happens because when we retrieve information from cache, we lost the reference that the variable is a class, and it becomes a regular JavaScript object. So the classToClass method on controllers doest not work. To fix this, we are now saving on cache already serialized.