Add cache cleaning when user updates profile
We had a problem where the provider didn't have the user profile change reflected in its appointment list because the cache was not being cleaned when the user updates the profile. Now that the cache is cleaned when any user updates its profile, a new query to the databse ensure that all information is up to date.