Change Input Autofill style and add new method to auth hook
The input had a problem where the autofill had no style. This was fixed adding style to the autofill hover.

Add a new method to auth hook to update the profile logged in the application. This was necessary because when the user updated its data or avatar, the changes were not being reflected.

A small change was made on Reset Password validation message.